o...k...
so. either nothing happens because we didn't name the function one of the things love recognizes. or the room is updating so fast that I can't actually see test popping up on the screen cuz the random(but not random) color updates to hide it too quickly.

need to come up with a way to turn off the rest of the game updating while we handle battles. can't just hide it.